noti update for update access of role

+++ app/controllers/v1/admin/iam/roles_controller.rb
@@ -54,7 +54,8 @@ class V1::Admin::Iam::RolesController < V1::ApplicationController
   # PATCH/PUT /v1/admin/iam/roles/:id
   def update
     if @role.update(role_params)
-      assign_permissions(@role) if permission_ids_param_provided?
+      permissions_changed = permission_ids_param_provided? && assign_permissions(@role)
+      notify_assigned_users(@role) if permissions_changed
 
       render_json_response(
         status_code: 200,
@@ -144,10 +145,22 @@ class V1::Admin::Iam::RolesController < V1::ApplicationController
   def assign_permissions(role)
     permission_ids = Array(permission_ids_param).reject(&:blank?)
     permissions = ::Iam::Permission.where(id: permission_ids)
+    current_permission_ids = role.permission_ids.map(&:to_s).sort
+    next_permission_ids = permissions.pluck(:id).map(&:to_s).sort
+
+    return false if current_permission_ids == next_permission_ids
 
     role.role_permissions.destroy_all
     permissions.each do |permission|
       role.role_permissions.find_or_create_by!(permission: permission)
+    end
+
+    true
+  end
+
+  def notify_assigned_users(role)
+    role.users.find_each do |user|
+      NotificationService::Center.iam_updated(user)
     end
   end
